moved Date to human readable output to one file reassembled po files
